classe EnderecoServiceImpl finalizada

+++ src/main/java/com/example/jbmotos/services/impl/EnderecoServiceImpl.java
@@ -1,14 +1,9 @@
 package com.example.jbmotos.services.impl;
 
 import com.example.jbmotos.api.dto.EnderecoDTO;
-import com.example.jbmotos.model.entity.Cliente;
 import com.example.jbmotos.model.entity.Endereco;
-import com.example.jbmotos.model.entity.Funcionario;
-import com.example.jbmotos.model.enums.TipoUsuario;
 import com.example.jbmotos.model.repositories.EnderecoRepository;
-import com.example.jbmotos.services.ClienteService;
 import com.example.jbmotos.services.EnderecoService;
-import com.example.jbmotos.services.FuncionarioService;
 import com.example.jbmotos.services.exception.ObjetoNaoEncontradoException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -26,18 +20,11 @@ public class EnderecoServiceImpl implements EnderecoService {
     private EnderecoRepository enderecoRepository;
 
     @Autowired
-    private ClienteService clienteService;
-
-    @Autowired
-    private FuncionarioService funcionarioService;
-
-    @Autowired
     private ModelMapper mapper;
 
     @Override
     @Transactional
     public Endereco salvarEndereco(EnderecoDTO enderecoDTO) {
-        verificaSeFuncionarioOuClienteExiste(enderecoDTO);
         return enderecoRepository.save(mapper.map(enderecoDTO, Endereco.class));
     }
 
@@ -50,33 +37,32 @@ public class EnderecoServiceImpl implements EnderecoService {
     @Override
     @Transactional(readOnly = true)
     public Optional<Endereco> buscarEnderecoPorId(Integer id) {
-        return enderecoRepository.findById(id);
+        Optional<Endereco> endereco = enderecoRepository.findById(id);
+        if (endereco.isEmpty()){
+            throw new ObjetoNaoEncontradoException("Endereço não encontrado para o Id informado.");
+        }
+        return endereco;
     }
 
     @Override
     @Transactional
     public Endereco atualizarEndereco(EnderecoDTO enderecoDTO) {
-        Objects.requireNonNull(enderecoDTO.getId(), "Erro ao tentar atualizar o Endereço. Informe um Id.");
+        Optional<Endereco> endereco = enderecoRepository.findById(enderecoDTO.getId());
+        if (endereco.isEmpty()){
+            throw new ObjetoNaoEncontradoException("Erro ao tentar atualizar." +
+                    " Endereço não encontrado para o Id informado.");
+        }
         return enderecoRepository.save(mapper.map(enderecoDTO, Endereco.class));
     }
 
     @Override
     @Transactional
-    public void deletarEndereco(Integer id) {
-        enderecoRepository.deleteById(id);
-    }
-
-    public void verificaSeFuncionarioOuClienteExiste(EnderecoDTO enderecoDTO) {
-        if (TipoUsuario.valueOf(enderecoDTO.getTipoUsuario()) == TipoUsuario.FUNCIONARIO) {
-            Optional<Funcionario> funcionario = funcionarioService
-                    .buscarFuncionarioPorCPF(enderecoDTO.getCpfUsuario());
-            funcionario.orElseThrow(() -> new ObjetoNaoEncontradoException("Erro ao tentar salvar o Endereço. " +
-                    "Funcionário não cadastrado, ou CPF inválido."));
-        } else if (TipoUsuario.valueOf(enderecoDTO.getTipoUsuario()) == TipoUsuario.CLIENTE) {
-            Optional<Cliente> cliente = clienteService.
-                    buscarClientePorCPF(enderecoDTO.getCpfUsuario());
-            cliente.orElseThrow(() -> new ObjetoNaoEncontradoException("Erro ao tentar salvar o Endereço. " +
-                    "Cliente não cadastrado, ou CPF inválido."));
+    public void deletarEnderecoPorId(Integer id) {
+        Optional<Endereco> endereco = enderecoRepository.findById(id);
+        if (endereco.isEmpty()){
+            throw new ObjetoNaoEncontradoException("Erro ao tentar deletar." +
+                    " Endereço não encontrado para o Id informado.");
         }
+        enderecoRepository.deleteById(id);
     }
 }
